version of the config in which there is an error with the require's
